+++ libs/string-utils/include/string_utils.h
@@ -1,6 +1,7 @@
 #ifndef STRING_UTILS_H_
 #define STRING_UTILS_H_
 
+#include <cstdint>
 #include <string>
 #include <vector>
 
@@ -9,7 +10,7 @@ namespace string_utils {
 std::vector<std::string> split_string(std::string str, char delim);
 
 // TODO: Generalize as template
-std::vector<int> numbers_from_string(std::string str);
+std::vector<int64_t> numbers_from_string(std::string str);
 
 }  // namespace string_utils
